requirements.txt
nltk
numpy
scikit-learn
transformers
tensorflow[and-cuda]
tf-keras
pillow
opencv-python
requests
librosa
python-magic
joblib
StrEnum
spacy
torch
requests
beautifulsoup4